Move link options to uv.gpr and make it possible to override libuv
Link options are now in uv.gpr, so users automatically link against
libuv.a or libuv.so.

Besides, the LIBUV_LINK_OPTION scenario variable makes it possible to
override which library to link against. This is useful at least for
developpers as they can link with a debug build of libuv instead of the
system one (in release mode: optimized and stripped).